refactor: Extract meta api to a module of its own
