+++ webpack/webpack.parts.js
@@ -5,6 +5,7 @@ const { WebpackPluginServe } = require("webpack-plugin-serve");
 const VirtualModulesPlugin = require("webpack-virtual-modules");
 const { extractStoriesJson } = require("./stories-json");
 const { importFn } = require("./importFn");
+const { attachMetaApi, setupFileWatcher } = require("./meta-api");
 
 const SKELETON_ENTRY = "./skeleton-entry.js";
 
@@ -114,7 +115,7 @@ const wps = ({ stories, configDir, projectDir }) => ({
           ctx.set("Access-Control-Allow-Origin", "*");
         });
 
-        addSideloadingAPI(app, storiesJson);
+        attachMetaApi(app, storiesJson);
         setupFileWatcher(projectDir);
       },
     }),
@@ -134,63 +135,11 @@ const wds = ({ stories, configDir, projectDir }) => ({
     after: async (app) => {
       const storiesJson = await extractStoriesJson({ stories, configDir });
 
-      addSideloadingAPI(app, storiesJson);
+      attachMetaApi(app, storiesJson);
       setupFileWatcher(projectDir);
     },
   },
 });
-
-// TODO: Handle updates (i.e. if something in glob changes).
-// Is there a hook for wds/wps for that?
-function addSideloadingAPI(app, storiesJson, projectDir) {
-  app.get("/api/ping", (_, res) => res.send("pong"));
-  app.get("/api/stories.json", (_, res) => res.json(storiesJson));
-  app.get("/api/stories/:id.json", ({ params: { id } }, res) => {
-    if (id) {
-      if (storiesJson.stories[id]) {
-        return res.json(storiesJson.stories[id]);
-      }
-
-      return res.sendStatus(404);
-    }
-
-    res.sendStatus(400);
-  });
-}
-
-// TODO: Connect the events with a websocket server
-function setupFileWatcher(projectDir) {
-  const Watchpack = require("watchpack");
-
-  // See https://www.npmjs.com/package/watchpack for full options.
-  // If you want less traffic, consider using aggregation with some interval
-  const wp = new Watchpack({
-    poll: true,
-    // Slower but the assumption is that some people use symlinks
-    followSymlinks: true,
-    ignored: "**/.git",
-  });
-
-  wp.watch({
-    directories: [projectDir],
-  });
-
-  wp.on("change", (filePath, mtime, explanation) => {
-    // filePath: the changed file
-    // mtime: last modified time for the changed file
-    // explanation: textual information how this change was detected
-    console.log("changed", { filePath, mtime, explanation });
-  });
-
-  wp.on("remove", (filePath, explanation) => {
-    // filePath: the removed file or directory
-    // explanation: textual information how this change was detected
-    console.log("removed", { filePath, explanation });
-  });
-
-  // TODO: This probably should get called when the server is being shut down
-  // wp.close();
-}
 
 const cpuProfiler = () => {
   const CpuProfilerWebpackPlugin = require("cpuprofile-webpack-plugin");
